refactor: GLOBAL_SETTINGS_MAPをglobal-settings.tsへ一元化
- executor.ts / validator.ts の二重定義（手動同期前提）を解消
- 重複していたloadGlobalSettings()も統合
- devに潜伏していたBiomeエラーを修正（未使用suppression 2件、
  lineWidth超過1行）。CIがdevで動いていなかったため未検出だった
